fix: 전역 상태 사용을 위한 파일 분리 및 정리
