--- App.style.js
@@ -1,13 +1,0 @@
-import { StyleSheet } from 'react-native';
-import { theme } from './theme';
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
-
-export default styles;
