+++ screen/index.js
@@ -4,5 +4,6 @@ import ListHome from './ListHome';
 import Search from './Search';
 import Setting from './Setting';
 import Login from './Login';
+import Home from './Home';
 
-export { Map, Populer, ListHome, Search, Setting, Login };
+export { Map, Populer, ListHome, Search, Setting, Login, Home };
